+++ src/main/java/io/github/taz/stickybot/listener/MessageContextInteractionListener.java
@@ -1,20 +1,17 @@
-package me.taz.stickybot.listener;
+package io.github.taz.stickybot.listener;
 
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.events.interaction.ModalInteractionEvent;
 import net.dv8tion.jda.api.events.interaction.command.MessageContextInteractionEvent;
-import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
-import net.dv8tion.jda.api.events.interaction.component.SelectMenuInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.requests.restaction.CommandListUpdateAction;
 
+import io.github.taz.stickybot.command.MessageCommand;
+import io.github.taz.stickybot.command.message.StickCommand;
+import io.github.taz.stickybot.command.message.UnStickCommand;
+
 import java.util.HashMap;
 import java.util.Map;
-
-import me.taz.stickybot.command.MessageCommand;
-import me.taz.stickybot.command.message.StickCommand;
-import me.taz.stickybot.command.message.UnStickCommand;
 
 public class MessageContextInteractionListener extends ListenerAdapter {
     public static final Map<String, MessageCommand> commandMap = new HashMap<>();
@@ -41,20 +38,5 @@ public class MessageContextInteractionListener extends ListenerAdapter {
     @Override
     public void onMessageContextInteraction(MessageContextInteractionEvent event) {
         commandMap.get(event.getName()).run(event);
-    }
-
-    @Override
-    public void onButtonInteraction(ButtonInteractionEvent event) {
-        commandMap.values().forEach(command -> command.onButtonInteraction(event));
-    }
-
-    @Override
-    public void onSelectMenuInteraction(SelectMenuInteractionEvent event) {
-        commandMap.values().forEach(command -> command.onSelectMenuInteraction(event));
-    }
-
-    @Override
-    public void onModalInteraction(ModalInteractionEvent event) {
-        commandMap.values().forEach(command -> command.onModalInteraction(event));
     }
 }
